test: added check_invariant to all the tests & relocated them
